Muzzle Flash Fix pt.2
The previous fix was incorrect. I was shooting slowly, and making the
muzzle flash static made it not happen when I was shooting slowly. When
shooting quickly, the muzzle flash still appeared. These changes include
various fixes that reduce the frequency of the muzzle flash.

For player convenience, I've added a line to move the muzzle flash off
the screen while it waits to garbage collect. This is *temporary*, while
I look into an actual fix.