Muzzle Flash Fix pt. 3
The Bullet class used the inherited destroy method. The bullet class was
responsible for cleaning up the muzzle flash, and it did so in it's
update loop when the muzzle flash finished it's animation. When the
bullet was removed from the autodrawset before the muzzle flash finished
it's animation, the muzzle flash stayed in the autodrawset permanently.

The fix: Override the Destroy method to include Muzzle.destroy(), and
the problem is resolved. The previous temporary fix has been removed.