Enemy XML fix
This fix includes changes to the FileParser, and removal of one method,
getEnemyData(). This method did not function correctly, and it's
intended behavior is covered by getEnemyData(EnemyType type). I've left
the removed method in a comment in the ConfigurationFileParser.java
file, with a comment title of "DEPRICATED".

What this fixes: This will fix the Portal Enemy error that was occurring
(causing it to fail the try/catch in it's constructor). This will also
fix the enemy speeds and sizes that were previously incorrect.

The issue: The XML file parser parses the XML data in the Enemies.xml
file line by line, and stores it in a map called enemies. This map is
made up of EnemyData objects, mapped by their EnemyType (such as
PORTAL_ENEMY). The original programmer left the EnemyData object they
were using as a global variable, that constantly got overwritten.
Therefore, all enemies shared the stats of the final enemy in the list.
Now the EnemyData object is no longer global, and enemies now hold their
correct data.

I will look into the other XML files to ensure that this isn't occurring
elsewhere in DyeHard, escaping our notice.