Spreadfire rectangle "fix"
The spreadfire weapon spawned a blue rectangle in the top corner when
activated. This was done in the constructor of the weapon, and appears
to be a left-over artifact of development. My guess is that an icon
indicating that you've equipped the gun should be in that corner. I've
removed the blue rectangle, as it servers no purpose as of now.

NOTE: The spreadfire gun appears to shoot three small bullets, and one
large bullet. The large bullet is due to called super.fire(), and
overlaps one of the smaller bullets. The front-shooting small bullet can
be removed from the fire method to make a more appealing shot, but I
didn't want to make that call.

NOTE2: The small bullets don't appear to make enemies change color, as
the large bullets do. However, because the Procedural API is diving into
collisions in a different way, I didn't want to spend time on fixing
this. Something to be aware of.