Refactor
This commit is an overhaul of the filestructure (put into
OO/Procedural), as well as an overhaul of all of the packages in the
project.

Previously, this was one large project with many packages. This commit
splits the repo into many projects, each with their own default package.
This includes a project for each API, of which all other projects use in
their build paths.

This commit also includes small renaming of some projects, to help
aleviate ambiguity and to maintain cohesiveness accross the
GameThemedGroup.