DH Tutorial Projects
Added the project files for the DyeHard tutorials-- these were missed in
the refactor. This includes stripping out packages, as before. Many of
the tutorials included cross-project dependencies (Tutorial 1 needed the
DebrisGenerator from Tutorial 5)-- these imports have been stripped, and
the class in question has been copied to each tutorial that needed it.